fix(distributed-locks): address PR #368 review findings
- restore ICanReceiveLockReleased as the wake-up DI seam (drop concrete-type injection)
- bound script SCRIPT LOAD under _loadScriptsLock with a linked-CTS timeout
- guard _eventsSubscribed with Interlocked so no handler leaks past Dispose
- semaphore TTL int->long (fix >24.8d overflow)
- extract DistributedLockHandleBase; unify lease-loss path with ownership probe
- read-only ZSET validate/count; prune only on acquire (not extend/read)
- collapse three *ScriptsInitializer into one RedisScriptsInitializer
- rename AddDistributedSemaphoreExtensions -> SetupDistributedSemaphore
- internal sealed DistributedSemaphoreProvider; drop redundant [PublicAPI]
- retain release-pipeline task on dispose timeout; reject InfiniteTimeSpan acquire
- correct NullDistributedSemaphoreProvider doc; add missing interface XML docs

+++ src/Headless.DistributedLocks.Redis/RedisDistributedSemaphoreStorage.cs
@@ -136,7 +136,7 @@ public sealed class RedisDistributedSemaphoreStorage(
             fenceKey,
             lockId,
             maxCount,
-            (int)ttl.TotalMilliseconds
+            (long)ttl.TotalMilliseconds
         );
         var result = await scriptsLoader
             .EvaluateAsync(Db, TryAcquireSemaphoreWithFenceScriptDefinition.Instance, parameters, cancellationToken)
@@ -163,7 +163,7 @@ public sealed class RedisDistributedSemaphoreStorage(
 
     private static SemaphoreSlotParams _GetSemaphoreSlotParameters(RedisKey holdersKey, string lockId, TimeSpan? ttl)
     {
-        var expiresValue = ttl.HasValue ? (int)ttl.Value.TotalMilliseconds : RedisValue.EmptyString;
+        var expiresValue = ttl.HasValue ? (long)ttl.Value.TotalMilliseconds : RedisValue.EmptyString;
 
         return new SemaphoreSlotParams(holdersKey, lockId, expiresValue);
     }
@@ -187,7 +187,7 @@ public sealed class RedisDistributedSemaphoreStorage(
         RedisKey fenceKey,
         string lockId,
         int maxCount,
-        int expires
+        long expires
     );
 
     [StructLayout(LayoutKind.Auto)]
